+++ src/giftCards/GiftCardCreateDialog/GiftCardCreateExpirySelect/index.tsx
@@ -1,0 +1,2 @@
+export * from "./GiftCardCreateExpirySelect";
+export { default } from "./GiftCardCreateExpirySelect";
