+++ src/giftCards/GiftCardCreateDialog/GiftCardCreateExpirySelect/messages.ts
@@ -1,0 +1,20 @@
+import { defineMessages } from "react-intl";
+
+export const giftCardCreateExpirySelectMessages = defineMessages({
+  expirySelectedLabel: {
+    defaultMessage: "Set gift card expiry date",
+    description: "GiftCarUpdateDetailsExpirySection expiry selected label"
+  },
+  expiryPeriodLabel: {
+    defaultMessage: "Expires in",
+    description: "GiftCarUpdateDetailsExpirySection expiry period label"
+  },
+  expiryDateLabel: {
+    defaultMessage: "Exact date",
+    description: "GiftCarUpdateDetailsExpirySection expiry date label"
+  },
+  expiryOnLabel: {
+    defaultMessage: "Will expire on:",
+    description: "GiftCarUpdateDetailsExpirySection expiry on label"
+  }
+});
